requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
sqlalchemy==1.4.49
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.10.13
httpx==0.23.3
python-multipart==0.0.6
